pyproject.toml
[tool.poetry]
name = "ai-finance-tracker"
version = "0.1.0"
description = "track your expenses and analyze the expenditure"
authors = ["chaksh <chakshita.shetty@spark.co.nz>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = ">=3.9"
fastapi = ">=0.115.12,<0.116.0"
uvicorn = ">=0.34.2,<0.35.0"
neo4j = ">=5.28.1,<6.0.0"
python-dotenv = ">=1.1.0,<2.0.0"
pandas = ">=2.2.3,<3.0.0"
python-multipart = ">=0.0.20,<0.0.21"
pdfplumber = ">=0.11.6,<0.12.0"
pypdf2 = "^3.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
